requirements.txt
streamlit
python-dotenv
PyPDF2
google-generativeai

langchain # core framework
langchain-huggingface # connect huggingface algorithms / models to perform embedding
faiss-cpu # Fast vector database to store the embedded data
langchain-community # extra integration 
langchain-text-splitters # split large texts into smaller chunks
sentence-transformers # call an pretrain embedding model to convert text into vectors
langchain-core # document, chains etc..
